don't fail if path doesn't exist

+++ src/java/com/rapleaf/cascading_ext/workflow2/TrashHelper.java
@@ -17,10 +17,12 @@ public class TrashHelper {
   }
 
   public static void deleteUsingTrashIfEnabled(FileSystem fs, Path path) throws IOException {
-    if(isEnabled()){
-      moveToTrash(fs, path);
-    }else{
-      fs.delete(path, true);
+    if(fs.exists(path)){
+      if(isEnabled()){
+        moveToTrash(fs, path);
+      }else{
+        fs.delete(path, true);
+      }
     }
   }
